% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes-package.R
\docType{package}
\name{quotes-package}
\alias{quotes-package}
\title{Collection of quotes for a variety of topics and from renowned personalities}
\description{
Random quotes directly from \code{R}! A collection of 61,071 unique quotes for a variety of topics and from renowned personalities. Useful for illustrating text mining analyses.
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

