% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes-package.R
\docType{package}
\name{quotes}
\alias{quotes}
\alias{quotes-package}
\title{Collection of quotes for a variety of topics and from renowned personalities}
\format{TODO}
\source{
TODO
}
\usage{
quotes
}
\description{
Random quotes directly from R! A collection of 75,966 quotes for a variety of topics and from renowned personalities, useful for text mining analyses.

TODO
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).

TODO
}
\keyword{datasets}

