% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquotes.R
\name{rquotes}
\alias{rquotes}
\title{Quote sampler}
\usage{
rquotes(n, author, genre, replace = TRUE)
}
\arguments{
\item{n}{TODO}

\item{author}{TODO}

\item{genre}{TODO}

\item{replace}{Should sampling be with replacement?}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{
set.seed(345678)
rquotes(n = 10)
rquotes(n = 10, genre = "science")
rquotes(n = 10, author = "Bertrand Russell")
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

