% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquotes.R
\name{rquotes}
\alias{rquotes}
\title{Sample random quotes}
\usage{
rquotes(n, auth, top, replace = FALSE)
}
\arguments{
\item{n}{number of quotes to sample.}

\item{auth}{filter by author, see examples.}

\item{top}{filter by topic, see examples.}

\item{replace}{Should sampling be with replacement?}
}
\value{
A \code{data.frame} formed by \code{n} rows of \code{\link{quotes}}.
}
\description{
Random selection of qutes, with optional author and topic filtering.
}
\details{
If
}
\examples{
# authors and topics
levels(quotes$author)
levels(quotes$topic)

# Random quotes
set.seed(345678)
rquotes(n = 10)
rquotes(n = 10, auth = "Bertrand Russell")
rquotes(n = 10, top = "science")
rquotes(n = 10, auth = "Bertrand Russell", top = "love")
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

